"check_style_score -> 0 of 1 passed. style score: 0/1 "

+++ cspp2-assignments/m4/Assignment-3/Solution.java
@@ -14,6 +14,7 @@ final class Solution
 	public static void main(final String[] args)
 	{
 		Scanner sc = new Scanner(System.in);
+		//sc.nextLine();
 		int n = Integer.parseInt(sc.nextLine());
 		for (int i = 0; i < n; i++) {
 			String s = sc.nextLine();
@@ -23,8 +24,18 @@ final class Solution
 				
 	}
 	public static long binaryToDecimal(String s) {
-	    int binary_value = Integer.parseInt(s);
-		int last_digit, i = 0;
+		//int binary = 1;
+		//int decimal = 0;
+		//for (int i =s.length()-1;i>=0;i--) {
+		//if(s.charAt(i) == '1') {
+		//decimal +=binary;
+		//}
+		//binary = binary *2;
+		//}
+		//String res = Integer.toString(decimal);
+		//return res;
+	    long binary_value = Integer.parseInt(s);
+		long last_digit, i = 0;
 		long decimal_value = 0;
 		while (binary_value != 0) {
 			last_digit = binary_value % 10;
